remove `Timestamps` and `Conflicts` from `FpcPayload`
