+++ bee-message/tests/fpc_payload.rs
@@ -3,7 +3,7 @@
 
 use bee_message::{
     payload::{
-        fpc::{Conflict, Conflicts, FpcPayload, Timestamp, Timestamps},
+        fpc::{Conflict, FpcPayload, Timestamp},
         transaction::TransactionId,
         MessagePayload,
     },
@@ -12,7 +12,7 @@ use bee_message::{
 use bee_packable::Packable;
 use bee_test::rand::bytes::rand_bytes_array;
 
-use std::{convert::TryInto, ops::Deref};
+use std::convert::TryInto;
 
 #[test]
 fn kind() {
@@ -27,7 +27,7 @@ fn version() {
 #[test]
 fn new_valid() {
     let fpc = FpcPayload::builder()
-        .with_conflicts(Conflicts::new(
+        .with_conflicts(
             vec![
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 0),
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 1),
@@ -35,8 +35,8 @@ fn new_valid() {
             ]
             .try_into()
             .unwrap(),
-        ))
-        .with_timestamps(Timestamps::new(
+        )
+        .with_timestamps(
             vec![
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 0),
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 1),
@@ -44,7 +44,7 @@ fn new_valid() {
             ]
             .try_into()
             .unwrap(),
-        ))
+        )
         .finish();
 
     assert!(fpc.is_ok());
@@ -72,37 +72,26 @@ fn timestamp_accessors_eq() {
 
 #[test]
 fn accessors_eq() {
-    let conflicts = Conflicts::new(
-        vec![
-            Conflict::new(TransactionId::from(rand_bytes_array()), 0, 0),
-            Conflict::new(TransactionId::from(rand_bytes_array()), 0, 1),
-            Conflict::new(TransactionId::from(rand_bytes_array()), 1, 2),
-        ]
-        .try_into()
-        .unwrap(),
-    );
+    let conflicts = vec![
+        Conflict::new(TransactionId::from(rand_bytes_array()), 0, 0),
+        Conflict::new(TransactionId::from(rand_bytes_array()), 0, 1),
+        Conflict::new(TransactionId::from(rand_bytes_array()), 1, 2),
+    ];
 
-    let timestamps = Timestamps::new(
-        vec![
-            Timestamp::new(MessageId::from(rand_bytes_array()), 0, 0),
-            Timestamp::new(MessageId::from(rand_bytes_array()), 0, 1),
-            Timestamp::new(MessageId::from(rand_bytes_array()), 1, 2),
-        ]
-        .try_into()
-        .unwrap(),
-    );
+    let timestamps = vec![
+        Timestamp::new(MessageId::from(rand_bytes_array()), 0, 0),
+        Timestamp::new(MessageId::from(rand_bytes_array()), 0, 1),
+        Timestamp::new(MessageId::from(rand_bytes_array()), 1, 2),
+    ];
 
     let fpc = FpcPayload::builder()
-        .with_conflicts(conflicts.clone())
-        .with_timestamps(timestamps.clone())
+        .with_conflicts(conflicts.clone().try_into().unwrap())
+        .with_timestamps(timestamps.clone().try_into().unwrap())
         .finish()
         .unwrap();
 
-    assert_eq!(fpc.conflicts().cloned().collect::<Vec<Conflict>>(), conflicts.deref());
-    assert_eq!(
-        fpc.timestamps().cloned().collect::<Vec<Timestamp>>(),
-        timestamps.deref()
-    );
+    assert_eq!(fpc.conflicts().cloned().collect::<Vec<Conflict>>(), conflicts);
+    assert_eq!(fpc.timestamps().cloned().collect::<Vec<Timestamp>>(), timestamps);
 }
 
 #[test]
@@ -130,9 +119,9 @@ fn unpack_valid() {
 }
 
 #[test]
-fn packable_round_trip() {
+fn round_trip() {
     let fpc_a = FpcPayload::builder()
-        .with_conflicts(Conflicts::new(
+        .with_conflicts(
             vec![
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 0),
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 1),
@@ -140,8 +129,8 @@ fn packable_round_trip() {
             ]
             .try_into()
             .unwrap(),
-        ))
-        .with_timestamps(Timestamps::new(
+        )
+        .with_timestamps(
             vec![
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 0),
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 1),
@@ -149,7 +138,7 @@ fn packable_round_trip() {
             ]
             .try_into()
             .unwrap(),
-        ))
+        )
         .finish()
         .unwrap();
 
@@ -161,7 +150,7 @@ fn packable_round_trip() {
 #[test]
 fn serde_round_trip() {
     let fpc_payload_1 = FpcPayload::builder()
-        .with_conflicts(Conflicts::new(
+        .with_conflicts(
             vec![
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 0),
                 Conflict::new(TransactionId::from(rand_bytes_array()), 0, 1),
@@ -169,8 +158,8 @@ fn serde_round_trip() {
             ]
             .try_into()
             .unwrap(),
-        ))
-        .with_timestamps(Timestamps::new(
+        )
+        .with_timestamps(
             vec![
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 0),
                 Timestamp::new(MessageId::from(rand_bytes_array()), 0, 1),
@@ -178,7 +167,7 @@ fn serde_round_trip() {
             ]
             .try_into()
             .unwrap(),
-        ))
+        )
         .finish()
         .unwrap();
     let json = serde_json::to_string(&fpc_payload_1).unwrap();
